Preserve cargo ashore and confirm losses before departure
Unload local cargo at landing, retain excess goods and mission freight across hull sales and reloads, and distribute freight before commodities at launch. Preview and confirm excess sales or job abandonment through the port UI. Keep player allegiance on every owned hull and use the authoritative credit balance in flight logs.

Validated 926 simulation tests, 33 engine tests, five runtime smokes, rendered port warnings, and a relocated Windows release. The save smoke now reloads excess cargo, cancels departure, then confirms the sale and launch.

+++ tests/sim/GauntletScenarios.cs
@@ -306,10 +306,10 @@ namespace EndlessSky.Tests
             player.Fleet.SetFlagship(hauler);
             player.SetCredits(1_000_000);
 
-            int capacity = player.Fleet.CargoCapacity();
+            long capacity = player.Fleet.CargoCapacity();
             Assert.Greater(capacity, 0, "a freighter has hold space");
 
-            int tons = Math.Min(capacity, 20);
+            int tons = (int)Math.Min(capacity, 20);
             long spend = (long)tons * cheapest.Price;
             player.AddCredits(-spend);
             int loaded = player.Fleet.LoadCargo(commodity, tons);
